Move env.handler to until
