--- app/helpers/env.handler.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-restricted-syntax */
-/* eslint-disable no-prototype-builtins */
-/* eslint-disable no-undef */
-
-exports.updateEnvVariables = (keys) => {
-  keys.map((key) => {
-    return delete process.env[key];
-  });
-};
