+++ app/util/env.handler.js
@@ -1,0 +1,5 @@
+exports.updateEnvVariables = (keys) => {
+  keys.map((key) => {
+    return delete process.env[key];
+  });
+};
